refactor(staticCloud): staticCloudCells 엔티티 삭제 및 밸류타입으로 리팩토링
