--- src/main/java/com/algangi/mongle/s2/repository/StaticCloudS2CellJpaRepository.java
@@ -1,8 +1,0 @@
-package com.algangi.mongle.s2.repository;
-
-import com.algangi.mongle.s2.domain.StaticCloudS2Cell;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StaticCloudS2CellJpaRepository extends JpaRepository<StaticCloudS2Cell, String> {
-
-}
