+++ src/main/java/com/algangi/mongle/staticCloud/domain/model/StaticCloud.java
@@ -1,10 +1,15 @@
-package com.algangi.mongle.staticCloud.domain;
+package com.algangi.mongle.staticCloud.domain.model;
 
+import java.util.Set;
+
+import jakarta.persistence.CollectionTable;
 import jakarta.persistence.Column;
+import jakarta.persistence.ElementCollection;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
 import jakarta.persistence.Table;
 import jakarta.persistence.UniqueConstraint;
 import lombok.AccessLevel;
@@ -40,11 +45,18 @@ public class StaticCloud {
     @Column(nullable = false)
     private Double longitude;
 
-    public static StaticCloud createStaticCloud(String name, Double latitude, Double longitude) {
+    @ElementCollection
+    @CollectionTable(name = "s2_static_cloud_cells",
+        joinColumns = @JoinColumn(name = "s2_token_id", unique = true)
+    )
+    private Set<String> s2TokenIds;
+
+    public static StaticCloud createStaticCloud(String name, Double latitude, Double longitude, Set<String> s2TokenIds) {
         return StaticCloud.builder()
             .name(name)
             .latitude(latitude)
             .longitude(longitude)
+            .s2TokenIds(s2TokenIds)
             .build();
     }
 }
