+++ src/main/java/com/algangi/mongle/staticCloud/repository/StaticCloudRepository.java
@@ -1,10 +1,12 @@
 package com.algangi.mongle.staticCloud.repository;
 
-import com.algangi.mongle.staticCloud.domain.StaticCloud;
 import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StaticCloudJpaRepository extends JpaRepository<StaticCloud, Long> {
+import com.algangi.mongle.staticCloud.domain.model.StaticCloud;
+
+public interface StaticCloudRepository extends JpaRepository<StaticCloud, Long> {
 
     Optional<StaticCloud> findByName(String name);
 }
